Move battle-stream-example down a level and pull out random-player-ai
