+++ sim/examples/random-player-ai.ts
@@ -2,28 +2,19 @@
  * Battle Stream Example
  * Pokemon Showdown - http://pokemonshowdown.com/
  *
- * Example of how to create AIs battling against each other.
+ * Example random player AI.
  *
  * @license MIT
  * @author Guangcong Luo <guangcongluo@gmail.com>
  */
 
-import * as BattleStreams from './battle-stream';
-import Dex = require('./dex');
-
-/*********************************************************************
- * Helper functions
- *********************************************************************/
+import {BattlePlayer} from '../battle-stream';
 
 function randomElem(array: number[]) {
 	return array[Math.floor(Math.random() * array.length)];
 }
 
-/*********************************************************************
- * Define AI
- *********************************************************************/
-
-class RandomPlayerAI extends BattleStreams.BattlePlayer {
+export class RandomPlayerAI extends BattlePlayer {
 	receiveRequest(request: AnyObject) {
 		if (request.wait) {
 			// wait request
@@ -67,39 +58,3 @@ class RandomPlayerAI extends BattleStreams.BattlePlayer {
 		}
 	}
 }
-
-/*********************************************************************
- * Run AI
- *********************************************************************/
-
-const streams = BattleStreams.getPlayerStreams(new BattleStreams.BattleStream());
-
-const spec = {
-	formatid: "gen7customgame",
-};
-const p1spec = {
-	name: "Bot 1",
-	team: Dex.packTeam(Dex.generateTeam('gen7randombattle')),
-};
-const p2spec = {
-	name: "Bot 2",
-	team: Dex.packTeam(Dex.generateTeam('gen7randombattle')),
-};
-
-const p1 = new RandomPlayerAI(streams.p1);
-const p2 = new RandomPlayerAI(streams.p2);
-
-console.log("p1 is " + p1.constructor.name);
-console.log("p2 is " + p2.constructor.name);
-
-(async () => {
-	let chunk;
-	// tslint:disable-next-line no-conditional-assignment
-	while ((chunk = await streams.omniscient.read())) {
-		console.log(chunk);
-	}
-})();
-
-streams.omniscient.write(`>start ${JSON.stringify(spec)}
->player p1 ${JSON.stringify(p1spec)}
->player p2 ${JSON.stringify(p2spec)}`);
